user service, dao refactor, security update
