+++ be/app/src/main/java/com/deckbop/api/service/AuthenticationService.java
@@ -3,7 +3,6 @@ package com.deckbop.api.service;
 
 import com.deckbop.api.controller.request.UserLoginRequest;
 import com.deckbop.api.exception.UserLoginException;
-import com.deckbop.api.model.User;
 import com.deckbop.api.security.jwt.JWTProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.BadCredentialsException;
@@ -14,7 +13,6 @@ import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import java.util.Map;
 import java.util.Optional;
 
 @Service
@@ -35,8 +33,8 @@ public class AuthenticationService {
         this.authenticationManagerBuilder = authenticationManagerBuilder;
     }
 
-    public String authorizeAndGetJWTToken(UserLoginRequest request) throws AuthenticationException, UserLoginException {
-        Optional<String> username = getUsernameFromCredentials(request.getCredentials());
+    public String authenticateAndGetJWTToken(UserLoginRequest request) throws AuthenticationException, UserLoginException {
+        Optional<String> username = userService.getUsernameFromCredentials(request.getCredentials());
         if (username.isEmpty()) {
             throw new UserLoginException("Bad credentials");
         }
@@ -53,26 +51,6 @@ public class AuthenticationService {
         SecurityContextHolder.getContext().setAuthentication(authentication);
         // boolean rememberMe = request.isRememberMe() != null && request.isRememberMe();
         return JWTProvider.createToken(authentication, true);
-    }
-
-    public Optional<String> getUsernameFromCredentials(Map<String, String> credentials) throws UserLoginException {
-        Optional<String> username = Optional.ofNullable(credentials.get("username"));
-        if (username.isEmpty()) {
-            Optional<String> email = Optional.ofNullable(credentials.get("email"));
-            if (email.isPresent()) {
-                Optional<User> user = userService.getUserByEmail(email.get());
-                if (user.isPresent()) {
-                    username = Optional.ofNullable(user.get().getUsername());
-                }
-                else {
-                    throw new UserLoginException("Invalid email");
-                }
-            }
-            else {
-                throw new UserLoginException("No credentials provided");
-            }
-        }
-        return username;
     }
 
 }
